requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
djangorestframework==3.13.1
greenlet==1.1.2
idna==3.3
pytz==2022.2.1
requests==2.28.1
SQLAlchemy==1.4.40
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
